backend/creo_ink/requirements.txt
wheel==0.40.0
django_channels==0.7.0
Django==4.2.3
djangorestframework==3.14.0
pylint-django==2.5.3
django-allauth==0.54.0
daphne==4.0.0
dj-rest-auth==4.0.1
djangorestframework_simplejwt==5.2.2
django-cors-headers==4.1.0

